QL: only make the toolbar sticky
