+++ tutor/src/screens/question-library/styles.scss
@@ -1,11 +1,5 @@
 @import '../screen-styles';
 
-.viewQuestionsLibrary {
-  padding-top: 0;
-  box-shadow: none;
-  padding: 0;
-}
-
 .questions-dashboard {
   $padding: 20px;
   .wrapper {
